商品添加抵用券

+++ src/Fx.Amiya.DbModels/DBModelConfigs/GoodsMemberRankPriceConfiguration.cs
@@ -14,7 +14,7 @@ namespace Fx.Amiya.DbModels.DBModelConfigs
             builder.ToTable("goods_member_rank_price");
             builder.Property(t => t.Id).HasColumnName("id").HasColumnType("varchar(50)").IsRequired();
             builder.Property(t => t.GoodsId).HasColumnName("goods_id").HasColumnType("varchar(50)").IsRequired();
-            builder.Property(t => t.MemberRankId).HasColumnName("member_rank_id").HasColumnType("int").IsRequired();
+            builder.Property(t => t.MemberRankId).HasColumnName("member_rank_id").HasColumnType("tinyint").IsRequired();
             builder.Property(t => t.Price).HasColumnName("price").HasColumnType("decimal(12,2)").IsRequired();
             builder.HasOne(e => e.GoodsInfo).WithMany(e => e.GoodsMemberRankPrice).HasForeignKey(e => e.GoodsId);
             builder.HasOne(e => e.MemberCardRankInfo).WithMany(e => e.GoodsMemberRankPrice).HasForeignKey(e => e.MemberRankId);
